RDKB-16428, RDKB-16307 : SNMP v3 queries are failing after maintanace window
Reason for change: snmp agent is been restarted in maintance window, which is killing snmp v3 subagent as well. Now we restarted the snmp restart only to snmpv2

Test Procedure:
1) After maintanance window make sure only snmp_subagent for snmpv2 is restarted and both v2 and snmv3 are functional by querying few parameters.
2) Kill snmpv2 subagent and check if it is restarts and check if both snmpv2 and snmpv3 are responding to queries.

Risks: Low

+++ scripts/run_subagent.sh
@@ -63,7 +63,9 @@ fi
 
 master=$1 # may empty
 
-killall snmp_subagent ; sleep 1
+if [ "$2" != "selfheal_snmpv2" ];then
+	killall snmp_subagent ; sleep 1
+fi
 
 if [ "x$master" != "x" ]; then
     echo "starting snmp_subagent process with $master"
